test,util(backend): add tag descriptions, try decimal in pydantic schemas
TODO: add conversion in router

+++ backend/src/ds_backend/http/schemas/company.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 from pydantic import BaseModel, Field
 from enum import Enum
+from decimal import Decimal
 
 # from ds_backend.http.schemas.stock_info import \
 #     StockInfoSchema
@@ -53,7 +54,7 @@ class CompanyInfoSchema(BaseModel):
         description="Price prediction. TODO: create model"
     )
 
-    stock_price: float = Field(
+    stock_price: Decimal = Field(
         description="Stock price in RUR, double decimal precision"
     )
 
@@ -61,7 +62,7 @@ class CompanyInfoSchema(BaseModel):
         description="Suggested stock action"
     )
 
-    prognosis_percentage: float = Field(
+    prognosis_percentage: Decimal = Field(
         description="Price prognosis for the stock"
     )
